Refactor lobby into modular packages
